# api-docs/api.yaml
openapi: 3.0.3
info:
    title: Blog API
    description: API documentation for a simple blog platform built with Express and MySQL.
    version: 1.0.0
servers:
    - url: http://localhost:4000
      description: Local development server

paths:
    /test:
        get:
            summary: Test the API
            description: Returns a simple JSON object to verify that the server is running.
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    lorem:
                                        type: string
                                        example: ipsum

    /blogposts:
        get:
            summary: Get all blogposts
            description: Returns a list of all blogposts and their authors.
            responses:
                '200':
                    description: A list of blogposts with authors
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Blogpost'
        post:
            summary: Create a new blogpost
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - title
                                - text
                            properties:
                                title:
                                    type: string
                                    example: How to Contribute to Open Source
                                summary:
                                    type: string
                                    example: A beginner-friendly guide to open source.
                                text:
                                    type: string
                                    example: Detailed steps on contributing to projects on GitHub.
            responses:
                '201':
                    description: Blogpost successfully created
                '422':
                    description: Missing required fields
                '500':
                    description: Server or database error

    /blogposts/{id}:
        get:
            summary: Get a blogpost by ID
            parameters:
                - in: path
                  name: id
                  schema:
                      type: integer
                  required: true
                  description: The ID of the blogpost to retrieve.
            responses:
                '200':
                    description: Blogpost found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Blogpost'
                '404':
                    description: Blogpost not found
        patch:
            summary: Update a blogpost
            description: Updates one or more fields of an existing blogpost.
            parameters:
                - in: path
                  name: id
                  schema:
                      type: integer
                  required: true
                  description: ID of the blogpost to update.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                    example: Updated title
                                summary:
                                    type: string
                                    example: Updated summary
                                text:
                                    type: string
                                    example: Updated full text content
            responses:
                '200':
                    description: Blogpost successfully updated
                '404':
                    description: Blogpost not found
                '422':
                    description: Missing or invalid fields
        delete:
            summary: Delete a blogpost
            parameters:
                - in: path
                  name: id
                  schema:
                      type: integer
                  required: true
                  description: ID of the blogpost to delete.
            responses:
                '200':
                    description: Blogpost deleted successfully
                '404':
                    description: Blogpost not found

components:
    schemas:
        Blogpost:
            type: object
            properties:
                id:
                    type: integer
                    example: 3
                title:
                    type: string
                    example: How to Contribute to Open Source
                summary:
                    type: string
                    example: A beginner-friendly guide to open source.
                text:
                    type: string
                    example: Step-by-step instructions for contributing to open source.
                date_posted:
                    type: string
                    format: date
                    example: 2025-10-31
                authors:
                    type: array
                    items:
                        $ref: '#/components/schemas/Author'

        Author:
            type: object
            properties:
                id:
                    type: integer
                    example: 1
                name:
                    type: string
                    example: Alice
                surname:
                    type: string
                    example: Johnson
                email:
                    type: string
                    format: email
                    example: alice@example.com